Updating branch with main files

+++ Controllers/BoardsController.php
@@ -4,8 +4,9 @@ namespace Controllers;
 
 use PDO;
 use Models\Board;
-use Models\List_app;
-use Models\User_app;
+// use Models\List_app;
+// use Models\User_app;
+use Models\Card;
 
 class BoardsController extends Controller
 {
@@ -14,7 +15,7 @@ class BoardsController extends Controller
      */
     public function __construct()
     {
-        // Vérifie si l'utilisateur est connecté sinon redirection
+        // Checks if the user is logged in otherwise redirects
         $this->redirectIfNotAuthenticated();
 
         parent::__construct();
@@ -25,22 +26,20 @@ class BoardsController extends Controller
      */
     public function add()
     {
-        $name_board = $_POST['nameboard'];
-        $color_board = $this->checkColor($_POST['color']);
-        $id_user_app = session('id');
- 
+        $name_board = trim($_POST['nameboard']);
+        $color_board = $this->checkColor($_POST['color']); //check if the color is acceptable
+        $id = session('id');
+
         $data = [
             'name_board' => $name_board,
             // 'cration_date_board' => /* date now*/,
             // 'position_board' => /* next position */,
             'color_board' => $color_board,
-            'id_user_app' => $id_user_app
+            'id_user_app' => $id
         ];
 
         $board = Board::createBoard($data);
-        // $id_board = Board::find();
-        // $id = Board::find($id);
-        return redirect('boards/show.php?id='. $board->id_board);
+        return redirect('/boards/show.php?id=' . $board->id);
     }
 
     /**
@@ -53,11 +52,41 @@ class BoardsController extends Controller
             return redirect('/home.php');
         }
 
+        // Security to check if the Board is updatable by user
+        
+        if (!Board::checkBoardOwnership($boardId, parent::getCurrentUserId())) {
+            messages("Nous n'avons pas trouvé cet élément.");
+            return redirect('/home.php');
+        }
+
         $board = Board::find($boardId);
         $lists = $board->listapp();
+
+        $listIds = [];
+        $newLists = [];
+
+        foreach ($lists as $list) {
+            $listIds[] = $list['id'];
+            $newLists[$list['id']] = $list;
+            $newLists[$list['id']]['cards'] = []; // newLists creates an empty array (that will content every card for that list) in all list's array. Each array matches the id of the list in it
+        }
+
+        $lists = $newLists; // $lists is an array that matches the id of list for convenience
+        unset($newLists);
+
+        $cards = Card::getCardsFromListIds($listIds);
+        foreach ($cards as $card) {
+            $lists[$card['id_list_app']]['cards'][] = $card; //fill every list'array with their cards
+        }
+        unset($cards);
+
         // Sort lists by position
-        // usort($lists, fn ($a, $b) => $a['position'] <=> $b['position']);
-        return $this->view('/boards/show.php', [
+        usort($lists, fn ($a, $b) => $a['position_list_app'] <=> $b['position_list_app']);
+        foreach ($lists as &$list) {
+            usort($list['cards'], fn ($a, $b) => $a['position_card'] <=> $b['position_card']);
+        }
+
+        return $this->view('boards/show.php', [
             'board' => $board,
             'lists' => $lists,
         ]);
@@ -68,19 +97,44 @@ class BoardsController extends Controller
      */
     public function update()
     {
-        if (empty($_POST['name'])){   
+
+        // Security to check if the board is updatable by user
+
+        $id = $_POST['id_board'];
+
+        if (!Board::checkBoardOwnership($id, parent::getCurrentUserId())) {
+            messages("Nous n'avons pas trouvé cet élément.");
+            return redirect('/home.php');
+        }
+
+
+        if (empty($_POST['name']) && empty($_POST['color'])) {
+            // $id = $_POST['id_board'];
+            return redirect('/boards/show.php?id=' . $id);
+        }
+
+        if (empty($_POST['name'])) {
             $color_board = $_POST['color'];
-            $id = $_POST['id_board'];
+            // $id = $_POST['id_board'];
             $data = [
                 'color_board' =>  $color_board,
                 'id' => $id
             ];
         }
 
-        if(!empty($_POST['name'])){
+        if (!empty($_POST['name'])) {
+            $name_board = $_POST['name'];
+            // $id = $_POST['id_board'];
+            $data = [
+                'name_board' => $name_board,
+                'id' => $id
+            ];
+        }
+
+        if (!empty($_POST['name'] && !empty($_POST['color']))) {
             $name_board = $_POST['name'];
             $color_board = $_POST['color'];
-            $id = $_POST['id_board'];
+            // $id = $_POST['id_board'];
             $data = [
                 'name_board' => $name_board,
                 'color_board' =>  $color_board,
@@ -88,7 +142,7 @@ class BoardsController extends Controller
             ];
         }
 
-        Board::updateBoard($data, $id); //$board = Board::update($boardId, $data);
+        Board::updateBoard($data, $id);
 
         return redirect('/boards/show.php?id=' . $id);
     }
@@ -101,7 +155,14 @@ class BoardsController extends Controller
         $boardId = $_POST['id_board'];
         $delete = $_POST['delete'];
 
-        if ($delete === "SUPPRIMER"){
+        // Security to check if the board can be deleted by this user
+
+        if (!Board::checkBoardOwnership($boardId, parent::getCurrentUserId())) {
+            messages("Nous n'avons pas trouvé cet élément.");
+            return redirect('/home.php');
+        }
+
+        if ($delete === "SUPPRIMER") {
             Board::delete($boardId);
             messages('Votre tableau a été supprimé avec succès.', 'alert-success');
             return redirect('/home.php');
@@ -112,21 +173,16 @@ class BoardsController extends Controller
     }
 
     /**
-     * @return string
+     * Check if the color value is correct to match a defined CSS theme
+     * @return null|void
      */
     public function checkColor($color)
     {
-        // if($color != 'blue' or $color != 'red' or $color != 'orange' or $color == null){
-        //     $color = 'blue';
-        //     return $color;
-
-        // }
-        // return $color;
-        if($color == 'red'){
+        if ($color == 'red') {
             return $color;
-        } else if ($color == 'blue'){
+        } else if ($color == 'blue') {
             return $color;
-        } else if ($color == 'orange'){
+        } else if ($color == 'orange') {
             return $color;
         } else {
             return $color = 'blue';
